requirements.txt
sqlalchemy
alembic
rich
pytest
fastapi[standard]
PyJWT
python-dotenv
psycopg2-binary
